created reports controller and relocated views files
