+++ Bangazon/Views/Shared/_Layout.cshtml
@@ -41,9 +41,8 @@
                         <li class="nav-item dropdown">
                             <a class="nav-link dropdown-toggle text-dark" href="#" id="navbarDropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Reports</a>
                             <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
-                                <a class="dropdown-item" asp-controller="Products" asp-action="IncompleteOrders">Incomplete Orders</a>
-                                <a class="dropdown-item" asp-controller="Products" asp-action="MultipleOrders">Multiple Orders</a>
-                                <a class="dropdown-item" asp-controller="Products" asp-action="AbandonedProductTypes">Abandoned Product Types</a>
+                                <a class="dropdown-item" asp-controller="Reports" asp-action="IncompleteOrders">Incomplete Orders</a>
+                                <a class="dropdown-item" asp-controller="Reports" asp-action="MultipleOrders">Multiple Orders</a>
                             </div>
                         </li>
                         <li class="nav-item">
